requirements.txt
requests
beautifulsoup4
feedparser
openai
python-pptx
python-dotenv
pandas
openpyxl
